Day 11 Context API, counter reborn.
This was painful, but in the end of the day I believe I understood it decently well.
the context itself is just an empty box, it should be placed in its own file, and initiated with createContext();
the context provider is a react component, that fills this box, and provides its values to the children that are wrapped in it. Initialized with <Context.Provider value={value you wanna pass, can be multiple}>
the consumer is any of the children wrapped in the context, or any of their children, and can request the context by using useContext hook.